changing bytes/string logic 180 degrees
analyzing how python 3's built-in wsgi server works, it feels
more sensible to keep data around as strings internally and only
on the very last moment convert it to bytes.

Also making sure the correct character set is specified in content-type

I'll need to make it possible for the handlers to specify the
encoding though, but doing utf-8 for now is the safest way to handle this.